flask-server/requirements.txt
tensorflow
numpy
pytype
flask>=2.2.2
flask-cors
pymongo
pillow